requirements.txt
numpy==1.24.3
flask
tensorflow==2.13.0
tensorflow_hub
roboflow
ultralytics==8.1.5
pillow

